Move ObjectStoragePath tests to sdk
