+++ task-sdk/tests/task_sdk/io/__init__.py
@@ -14,10 +14,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
-from __future__ import annotations
-
-from airflow.sdk.io.path import ObjectStoragePath
-from airflow.sdk.io.store import attach
-
-__all__ = ["ObjectStoragePath", "attach"]
